refactor: Move all files under java instead of under zephyr
