+++ src/main/java/Zephyr.java
@@ -1,0 +1,75 @@
+package zephyr;
+
+import java.io.IOException;
+
+import zephyr.Commands.AbstractCommand;
+import zephyr.Controllers.Parser;
+import zephyr.Controllers.Storage;
+import zephyr.Controllers.Ui;
+import zephyr.DataStructures.TaskList;
+import zephyr.Exceptions.ZephyrException;
+
+/**
+ * The main class for the Zephyr application.
+ * This class is responsible for initializing the user interface, storage, parser,
+ * and task list, as well as running the main application loop.
+ */
+public class Zephyr {
+
+    private final Ui ui;
+    private TaskList tasks;
+    private final Storage storage;
+    private final Parser parser;
+
+    /**
+     * Constructs a Zephyr application with the specified file path for task storage.
+     *
+     * @param filePath the file path where tasks are stored in markdown format
+     */
+    public Zephyr(String filePath) {
+        ui = new Ui();
+        storage = new Storage(filePath);
+        parser = new Parser(storage, ui);
+        try {
+            tasks = new TaskList(storage.loadFile());
+        } catch (IOException e) {
+            ui.printLoadingError();
+            tasks = new TaskList();
+        }
+    }
+
+    /**
+     * Runs the main application loop.
+     * This method prints the welcome message, then enters a loop to read user commands,
+     * execute them, and check for an exit command. When an exit command is received,
+     * the loop terminates and the tasks are saved to storage.
+     */
+    public void run() {
+        ui.printWelcome();
+        boolean isExit = false;
+        while (!isExit) {
+            try {
+                String fullCommand = ui.readCommand();
+                AbstractCommand c = parser.parse(fullCommand);
+                c.execute(tasks, ui, storage);
+                isExit = c.isExit();
+            } catch (ZephyrException e) {
+                ui.println(e.getMessage());
+            }
+        }
+        try {
+            storage.saveFile(tasks.getTasks());
+        } catch (IOException e) {
+            ui.printSavingError();
+        }
+    }
+
+    /**
+     * The main entry point of the Zephyr application.
+     *
+     * @param args command-line arguments (not used)
+     */
+    public static void main(String[] args) {
+        new Zephyr("data/tasks.md").run();
+    }
+}
