+++ src/main/java/commands/FindCommand.java
@@ -1,0 +1,54 @@
+package zephyr.Commands;
+
+import java.util.List;
+import zephyr.Controllers.Storage;
+import zephyr.Controllers.Ui;
+import zephyr.DataStructures.TaskList;
+import zephyr.Exceptions.ZephyrException;
+import zephyr.Tasks.AbstractTask;
+
+/**
+ * Represents a command to search for tasks based on a provided query.
+ * The command filters the tasks in the TaskList whose descriptions contain
+ * the search query and prints the matching tasks.
+ */
+public class FindCommand extends AbstractCommand {
+
+    /**
+     * Constructs a FindCommand with the specified search query.
+     *
+     * @param arguments the search query used to filter tasks
+     */
+    public FindCommand(String arguments) {
+        super(arguments);
+    }
+
+    /**
+     * Executes the find command by searching for tasks that match the query.
+     * This method retrieves tasks from the TaskList whose descriptions contain
+     * the search query and then prints the found tasks using the Ui object.
+     *
+     * @param tasks   the TaskList to search within
+     * @param ui      the Ui used to display the search results
+     * @param storage the Storage object (not used in this command)
+     * @throws ZephyrException if an error occurs during the search execution
+     */
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) throws ZephyrException {
+        List<AbstractTask> foundTasks = tasks.findTasks(arguments);
+        ui.printTaskList(foundTasks);
+    }
+
+    /**
+     * Validates the find command arguments.
+     * This method ensures that a non-empty search query is provided.
+     *
+     * @throws ZephyrException if the search query is empty
+     */
+    @Override
+    public void isValidCommand() throws ZephyrException {
+        if (arguments.isEmpty()) {
+            throw new ZephyrException("The description of a find command cannot be empty.");
+        }
+    }
+}
