+++ src/main/java/exceptions/ZephyrException.java
@@ -1,0 +1,15 @@
+package zephyr.Exceptions;
+
+/**
+ * A custom runtime exception for handling errors in the Zephyr application.
+ */
+public class ZephyrException extends RuntimeException {
+    /**
+     * Constructs a new ZephyrException with the specified detail message.
+     *
+     * @param message the detail message
+     */
+    public ZephyrException(String message) {
+        super(message);
+    }
+}
